add "patreon" (share), profile string, uint32 days
Added income sharing ("Patreon") functionality through the setshare and resetshare actions.

Added an entirely optional "profile" string to be attached to a token balance that can have up to 1024 bytes. The idea is that this would allow a very basic "social networking" feature to be bootstrapped.

Widened the last-claim-day counter from 16 bits to 32 bits, to avoid having to write wrap-around code.